requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord.py==2.0.1
docopt==0.6.2
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
pipreqs==0.4.11
pycodestyle==2.9.1
python-dotenv==0.21.0
requests==2.28.1
toml==0.10.2
urllib3==1.26.12
yarg==0.1.9
yarl==1.8.1
